requirements.txt
fastapi==0.109.2
langchain-google-genai
google-generativeai
langchain-community==0.0.34
faiss-cpu==1.8.0
python-multipart==0.0.9
httpx==0.27.0
langchain 
qdrant-client 
python-dotenv 
langchain-google-genai
flask-cors
uvicorn